requirements.txt
google-cloud-aiplatform
google-auth
fastapi
pydantic
uvicorn
python-dotenv